sasl: support tls-exporter channel binding
RFC 9266 was written to allow channel binding with TLS 1.3 (for which
previous channel binding methods are not defined).
This patch supports the new tls-exporter channel binding type when the
TLS version in use is 1.3 or greater, and falls back to tls-unique
otherwise (and only if it is available).

Fixes mellium/xmpp#136

Signed-off-by: Sam Whited <sam@samwhited.com>