sasl: add Nonce option for use in testing
Previously the testing package needed to be able to access internal
state details of the sasl package to make the output deterministic.
This patch adds a "Nonce" option that is only exported in tests that
allows setting the random nonce to anything, thereby obviating the need
for full access to internal state.
It also changes the testing packgae to be the sasl_test package so that
internal state can't be used in the future.

Signed-off-by: Sam Whited <sam@samwhited.com>